fix: change SubscribeResponse (#19)

+++ src/main/java/back/whats_your_ETF/dto/SubscribeResponse.java
@@ -3,5 +3,5 @@ package back.whats_your_ETF.dto;
 public record SubscribeResponse(
     Long publisher_id,
     String nickname,
-    Long revenue
+    Double revenue
 ){}
